first commit
Please attention, the table that was tried to be connected was created in XHINCE scheme, not in DUKKAN scheme for easy trial-changing. Therefore, not every user who tries to reach may have authority or permission. Please create the same table structures in your personal scheme and log in with your own user / pass. (application.properties)